generalize element tuple rule
